tree table ts [Improvement] (#213)
* jsx to tsx, tree table componen.

* Typing.

* Typing.

* Typing.

* Typing.

* Typing.

* Typing.

* Typing.

* Typing.

* Typing.

* Code formatting.

* Code formatting.

* Typing.

* Typing.

* Typing.

* Typing.

* Typing

* Linting

* cleanup

* Typing.

* Typing.

* Delete file.

* Linting.

* Code formatingg.
