+++ pyrene/package.json
@@ -48,6 +48,7 @@
     "@types/jest": "^26.0.23",
     "@types/react": "^16.9.50",
     "@types/react-select": "^3.0.28",
+    "@types/react-window": "^1.8.5",
     "@typescript-eslint/eslint-plugin": "^4.28.1",
     "@typescript-eslint/parser": "^4.28.1",
     "autoprefixer": "^10.2.6",
